+++ package.json
@@ -5,11 +5,14 @@
   "main": "index.js",
   "dependencies": {
     "body-parser": "^1.19.0",
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "webpack": "^4.41.0",
+    "webpack-cli": "^3.3.9"
   },
   "devDependencies": {},
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "webpack",
+    "test": "jasmine"
   },
   "repository": {
     "type": "git",
