+++ webpack.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+	entry: './script/script',
+	output: {
+		filename: './build.js'
+	}
+}
